backend/pyproject.toml
[tool.poetry]
package-mode = false
name = "backend"
version = "0.1.0"
description = ""
authors = ["Urii Alexeev <slmpnv.code.mail@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
psycopg = "^3.2.3"
pydantic = "^2.10.3"
python-dotenv = "^1.0.1"
icecream = "^2.1.3"
redis = "^5.2.1"
ipykernel = "^6.29.5"
psycopg-binary = "^3.2.6"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
